New font-system, window-icon, cleanup
Changed how fonts are made\used, so that it's easy to draw-strings of varying sizes, and not need to create a separate font every-time (Instead, multiple font-objects get created in a variety of sizes initially instead).

Created separate font-path and asset-path for default loading of them. font-path defaults to asset-path.

made it possible to set an icon for the window.

Preparing to make it possible to pause animations (not implemented).