updated image-drawing, and better init step
Made the :init be within the main-loop instead, makes it easy to have loading-screens etc for asset-initialization.

Fixed up drawing and auto-drawing to be more consistent. You can now either use image-name or an image-object when using\refering to images as it pertains to drawing, or auto-drawing.

Will fix up collision to also work the same (where you can refer to images with names, rather than the object itself)