changes to animation, new example and fixed an issue with asset-path
Fixed up flipping to iron out some minor kinks.

Fixed asset path so that it won't throw an error if you don't specify a path when ASDF system and package doesn't match (it'll default to empty string).

made animations work with auto-drawing.

added an simple showcase (showing rectangles) to give an quick and easy perspective of how to use the engine, contra animated-sprite which is pretty noicy.

removed the gui.lisp as I'll need to do some serious rethinking on how to implement something like that, very low priority, so no point in having it around.